- lexer.rl: make sure all invalid escapes lead to fatal errors (fixes #136).
Due to the way escape sequences consume the following characters,
it does not make sense to attempt to continue lexing after
encountering an error. Additionally, if diagnostics consumer will not
raise, lexing code will attempt to append Proc in @escape to
the buffer. Rather than attempting to define a fallback value for
escape sequences, it is made to always be fatal.