* ruby20.y: revert 7f7f2a45.
Ruby bug https://bugs.ruby-lang.org/issues/11107 stated that
the change will be backported to 2.0.0, but it is not, and 2.0.0
support will be stopped at 25 Feb 2015 (in 8 months), so it is
unlikely to ever be.

Thus, the change and corresponding bump of the syntax deviation
warning is reverted to avoid stalling further Parser releases.