Rewriter merges actions when possible, avoiding unnecessary 'clobber' errors
If 2 rewrite actions both attempt to delete the same piece of text, there is
no reason to raise a 'clobber' exception. Likewise, if they both replace the
same text with the same replacement, that is fine.

Even if they replace 2 overlapping regions of text, that is still fine IF
the overlapping portion of the replacement is the same.