- lexer.rl: "f(a ? 'a':1)": disallow quoted label at expr_value.
This commit tracks upstream commits ruby/ruby@94b1db5d and
ruby/ruby@0958af2ad and Ruby bug https://bugs.ruby-lang.org/issues/10653.

You may notice that my implementation is much simpler. This is because
I'm lazy (and also the bitfield hack wouldn't really work on Ragel).

In short, the MRI implementation seems to give up on enumerating
possible contexts and instead now explicitly tracks where a label,
quoted or not, can appear.

My implementation is based on the observation that the crux of
the issue is that EXPR_VALUE affects regular labels (e.g. `a:`),
forbidding them in certain contexts, but does not affect quoted
labels (e.g. `"a":`) in the same way. Therefore, the only necessary
change is to extend the existing semantics of EXPR_VALUE.

It seems sound to me, and all tests pass, but it seems slightly odd
that a trivial 40 line patch could do the same work as a 340 line
patch that overhauls a large part of lexer and adds two more parser
hacks.

I don't really know.