Simplify lexer code which unnecessarily uses emit_table
In a number of cases, a table lookup is used to find the token to emit, although
there is only one possible token which could be emitted from that point. It
makes the code clearer when the reader can see the token which will be emitted,
rather than having to go and consult the table.