?\u{9} should not match error rule
The rule which throws an error for invalid \u escapes would match \u{9}.
This was never noticed, because rather than directly raising an error,
it would assign a lambda to @escape, which would subsequently be overwritten
before @escape was actually used.

However, it's better if the error rule works exactly as required -- in the
future, rework of the code could cause the error to show itself.