Add new lexer state especially for backslash-delimited strings
The previous code didn't handle backslash-delimited strings correctly in all
cases (specifically, it failed when the ending string delimiter was right at
EOF).

It didn't have any test coverage either, and it forced us to store error lambdas
in @escape rather than directly raising an error when we lex an invalid escape.
That made the code convoluted and harder to understand.