Source::Rewriter handles clobbering correctly for insertions
Previously, insertions ("replacements" of a zero-length range) would never
clobber or be clobbered by anything, even if they landed right in the middle
of a replaced or deleted range of text. As can easily be seen, this makes a
mess and should give rise to a clobbering error.

The tricky thing is that an insertion can legitimately appear at the very
beginning or very end of a replaced or deleted range. If it is in the middle,
though, that is definitely wrong and should cause an error. (Unless it appears
in the middle of a replaced range, and the replacement is longer than the
replaced text by the length of the insertion or more, and the replacement has
the same effect as the insertion.)

Additionally, two insertions at the very same point should also clobber each
other. Why? Well, how do we know which inserted text should come before the
other? We don't. And if the inserted texts are identical, how do we know if
the intention was to add one or two copies of that text? Again, we don't. It's
safer to just flag this as clobbering.

Additionally, there were some cases where the previous code would merge edits
which should really have raised a clobbering error. It is now more robust all
around, and close to 30 new unit tests have been added to verify its behavior.