Make Buffer#line_begins fly
The existing code uses `String#each_char`, which allocates a new string object
for EACH character in the source file. (This is the #2 source of object
allocations in RuboCop.) It also uses a lot of (slow) Ruby method calls.

Switching to `String#index` increases the performance of
`Source::Buffer#line_begins` by a factor of more than 20.