In Source::Comment::Associator, don't traverse subnodes with no comments
If the location of the next comment is completely beyond the parent node, then
we know that none of the subnodes will be associated with a comment. So skip
over them, rather than traversing them for nothing.

There is one type of node which can have children outside its own 'expression'
range; a heredoc. So we always need to traverse the child nodes of a heredoc.