Generalize the unicode performance fixes.
This is an attempt to generalize the performance gains of #269 to extend
to clients of the resulting AST.

The immediate problem is that many AST consumers (rubocop is the one I'm
currently interested in) are going to invoke
`Parser::Source::Range#source` on many of the AST nodes, and that method
has the same linear-time-slicing behavior as `Parser::Lexer#tok` did.

By pushing the re-encoding inside the `Parser::Source::Buffer`
abstraction, we make it also available to `Range`, and we simplify
`Lexer` while we're at it: A bunch of encoding bookkeeping is now hidden
inside the `Buffer#slice` abstraction.