- Builders::Default: reject non-UTF-8 compatible literals.
Downstream tooling does not want to deal with literals that arrive
with invalid encoding in exceedingly rare edge cases. Such cases are
hard to handle and tedious to test, with little, if any, benefit.

We already reject non-ASCII-compatible encodings outright, so this
is not a substantial departure from existing practice. In any case,
if downstream tooling wishes to handle this, it can inherit from
Builders::Default and override the corresponding methods.

Fixes #252.