ruby24: Generate proper AST for multi-asgn in condition
Commit c1310f6 introduces a new behavior that got merged for the Ruby
2.4 release: multi-asgn is now allowed in a conditional context.

Besides the changes to the language grammar, the commit also updates the
AST builder (builder/default.rb @ check_condition) to stop it from
emiting a diagnostic error in 2.4.

This change, however, is not correct: the condition check for Ruby 2.4
does not emit an error anymore, but it does not return the AST nodes it
was checking either. Without an explicit return, the `:masgn` branch of
the function yields a `nil`, swallowing up the expressions inside the
condition of the `if` statement.

The fix is trivial: return the value we were checking. The corresponding
test has been updated to reflect the change -- the previous AST output
was missing all the contents of the `if` statement.