Fix performance regression for a string that contains a very long line.
`+=` is replaced with `<<` in `extend_string` method in #293.
This replacing did not introduce performance regressions about small string. However, it has introduced a performance regression about a string that contains very long line.

This seems to be a problem in an actual application. An issue was opened
in RuboCop repository about this problem. https://github.com/bbatsov/rubocop/issues/4701

`<<` is replaced with `+=` also in `initialize` method by #293. However this `<<` doesn't introduce regressions because it is a concatenation of small strings.

Since #293 was added to support Opal, and Opal differs from any other Ruby implementation
in its support for mutable strings, the troublesome parts of #293 are reverted.