Don't add dozens of encoding comments to each generated file
The 'generate' Rake task runs every time the tests are run, YARD documentation is
generated, and so on. Even if all the generated files are already there, it adds another
encoding comment to them each time. If you repeatedly run the tests while working on the
Parser codebase, these generated files accumulate more and more encoding comments at the
top.

Put an end to this silliness by only adding the encoding comment if it is not already there.