* Bump maintenance branches to 3.1.2, 3.0.4, 2.7.6, and 2.6.10
These Rubies has been released.

- https://www.ruby-lang.org/en/news/2022/04/12/ruby-3-1-2-released/
- https://www.ruby-lang.org/en/news/2022/04/12/ruby-3-0-4-released/
- https://www.ruby-lang.org/en/news/2022/04/12/ruby-2-7-6-released/
- https://www.ruby-lang.org/en/news/2022/04/12/ruby-2-6-10-released/

## Ruby 3.1 branch

Bump 3.1 branch from 3.1.1 to 3.1.2

https://github.com/ruby/ruby/compare/v3_1_1...v3_1_2

There seems to be no change to the syntax.

## Ruby 3.0 branch

Bump 3.0 branch from 3.0.3 to 3.0.4

https://github.com/ruby/ruby/compare/v3_0_3...v3_0_4

There seems to be no change to the syntax.

## Ruby 2.7 branch

Bump 2.7 branch from 2.7.5 to 2.7.6

https://github.com/ruby/ruby/compare/v2_7_5...v2_7_6

There seems to be no change to the syntax.

## Ruby 2.6 branch

Bump 2.6 branch from 2.6.9 to 2.6.10

https://github.com/ruby/ruby/compare/v2_6_9...v2_6_10

There seems to be no change to the syntax.