add tests for newline after hash key in pattern matching.
This commit tracks upstream commit ruby/ruby@eaeb130.

Ruby < 3.2 had a bug that emitted tNL if there's a newline
after hash key in pattern matching like

```ruby
case foo
in a:
  0
  1
end
```

Before 3.2 it was `in a: 1` pattern, starting from 3.2 it's `in a:` with a multiline body.

Unfortunately unsetting `in_kwarg` in parsers < 3.2 is not enough to handle all cases,
tNL must still be emitted in a few cases, for example for command:

```ruby
in a:
  p a
```

but not if it's a trivial expression. I think backporting this bug is not worth it,
the chance of introduce new bugs is too high. If you know what is a proper fix
please don't hesitate to send a PR.

This PR only adds tests for new behaviour that is implemented
by all versions of parser.