Outline numeric literal int digit handling
This code was inlined 12 times in the F1 lexer. Outlining it
reduces advance instruction count from 16879 to 15807 on JRuby.

Note that the bottom section of this part depends on the @num_xfrm
closure being able to see and update the local `p`, which prevents
outlining it from the main body.