* Bump maintenance branches to 3.2.2, 3.1.4, 3.0.6, and 2.7.8
These Rubies have been released.

- https://www.ruby-lang.org/en/news/2023/03/30/ruby-3-2-2-released/
- https://www.ruby-lang.org/en/news/2023/03/30/ruby-3-1-4-released/
- https://www.ruby-lang.org/en/news/2023/03/30/ruby-3-0-6-released/
- https://www.ruby-lang.org/en/news/2023/03/30/ruby-2-7-8-released/

## Ruby 3.2 branch

Bump 3.2 branch from 3.2.1 to 3.2.2

https://github.com/ruby/ruby/compare/v3_2_1...v3_2_2

There seems to be no change to syntax.

## Ruby 3.1 branch

Bump 3.1 branch from 3.1.3 to 3.1.4

https://github.com/ruby/ruby/compare/v3_1_3...v3_1_4

There seems to be no change to syntax.

## Ruby 3.0 branch

Bump 3.0 branch from 3.0.5 to 3.0.6

https://github.com/ruby/ruby/compare/v3_0_5...v3_0_6

There seems to be no change to syntax.

## Ruby 2.7 branch

Bump 2.7 branch from 2.7.7 to 2.7.8

https://github.com/ruby/ruby/compare/v2_7_7...v2_7_8

There seems to be no change to syntax.