+ add prism-specific node types (#1071)
It's a bit awkward: rubocop and rubocop-ast use this list to generate visitors and node type predicates.
Since they both use this, it doesn't make much sense to add this handling in rubocop I think.

On the other hand, adding this to prism doesn't really work since the parser
there is lazily loaded, and by rubocop as well since users may just use older ruby versions.
Prism itself has no dependecy on parser to eagerly add to this list.

I think adding this to parser is ok, and is also the simplest solution to the problem.
The list can (and does) contain entries that are parser-version specific, this just
extends that policy a bit farther than it used to be.